wip rewrite: read host archive instead of export format
It would be nice if it were possible to read the export directly,
but there are some issues which make that impractical:

- the export format is poorly documented and poorly designed.
- the only library support for that format is in javascript.
- even if we write our own parser for the format, we'd then
  have to write a renderer which could re-emit it as markdown.
- even if we did all that, the export format doesn't contain
  any of the site's images.

All told, it's much better to unpack a complete archive of
the ghost site, read the DB directly, extract the markdown
source and metadata from there, and unpack appropriate images.

This is very WIP; there's a lot left to do. That said, `find_ghost_db`
is already beating `tar -tjf | rg 'ghost\.db$'` by about 10% (avg.
after 1 run), which is nice.